Have a default path for quantum-rootwrap
Finding the path to quantum-rootwrap in a ruby_block makes
node[:quantum][:rootwrap] exist too late to be used when populating
quantum.conf in the first chef-client run. This results in some failures
to use sudo, which are not critical but annoying.

Instead of proceeding this way we just default the path to
quantum-rootwrap to /usr/bin/quantum-rootwrap (correct path for
upstream, Ubuntu and SUSE), and keep the ruby_block to fix the path if
it's not the same. Things will still fail in the first chef-client run
if it's not the right path, but chef convergence will make it work for
later runs. And it now works directly in the first run in most cases.